[#163964746] Structure app to meet ADC folder structure
