+++ UI/js/common_scripts.js
