--- UI/layout.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<title>Politico</title>
-	<link rel="stylesheet" type="text/css" href="../public/css/style.css">
-</head>
-<body>
-	<nav class="navbar">
-		<span class="navbar-toggle" id="js-navbar-toggle">
-			<i class="fas fa-bars">Menu</i>
-		</span>
-		<a href="#" class="logo">logo</a>
-		<ul class="main-nav" id="js-menu">
-		</ul>
-	</nav>
-	<section class="row">
-		<h2 class="text-center">List of people</h2>
-	</section>
-	<script type="text/javascript" src="../public/navselect.js"></script>
-</body>
-</html>
