--- UI/partials/nav.html
@@ -1,8 +1,0 @@
-<nav class="navbar">
-	<span class="navbar-toggle" id="js-navbar-toggle">
-		<i class="fas fa-bars">Menu</i>
-	</span>
-	<a href="#" class="logo">logo</a>
-	<ul class="main-nav" id="js-menu">
-	</ul>
-</nav>
