--- UI/recoverpwd.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="UTF-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<title>Sign in @ Politico</title>
-		<link rel="stylesheet" type="text/css" href="../public/css/style.css">
-	</head>
-	<body>
-		<div class="navbar">
-			<h2 class="text-center">Politico</h2>
-		</div>
-		<div class="row">
-			<h3 class="text-center">Recover the password</h3>
-		</div>
-		<div class="row">
-			<div class="col-smt-12 col-tab-2 col-lap-3 col-dsk-3"></div>
-			<div class="col-smt-12 col-tab-8 col-lap-6 col-dsk-6">
-				<label for="email">Email</label>
-				<input type="email" name="email" id="email" placeholder="Provide your email address">
-				<button class="btn btn-save" onclick="getAuth()">Recover</button>
-				<div class="col-smt-12 gray-area">
-					<h4>Or <a href="signin.html">sign in</a></h4>
-				</div>
-			</div>
-			<div class="col-smt-12 col-tab-2 col-lap-3 col-dsk-3"></div>
-		</div>
-		<script type="text/javascript" src="../public/js/http_request_services.js"></script>
-		<script type="text/javascript" src="../public/js/notifier.js"></script>
-		<script type="text/javascript" src="../public/js/js_auth.js"></script>
-	</body>
-</html>
