--- controllers/pagesController.js
@@ -1,8 +1,0 @@
-const randomstring = require('randomstring');
-
-exports.getMainPage = (req, res) => {
-	return res.render('parties', {
-		title: 'Parties and offices',
-		rString: randomstring.generate(9),
-	});
-};
