--- dataHelper/data/offices.json
@@ -1,1 +1,0 @@
-[{"id":6,"name":"Office Test","type":"federal","createdAt":"Wed Feb 13 2019 12:08:29 GMT+0200 (Central Africa Time)","updatedAt":"Wed Feb 13 2019 12:08:29 GMT+0200 (Central Africa Time)"}]
