--- dataHelper/data/parties.json
@@ -1,1 +1,0 @@
-[{"id":6,"name":"Part Test","hqAddress":"Kigali","logoUrl":"testUrl","createdAt":"Wed Feb 13 2019 12:08:29 GMT+0200 (Central Africa Time)","updatedAt":"Wed Feb 13 2019 12:08:29 GMT+0200 (Central Africa Time)"}]
