--- index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="UTF-8">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<title>Sign in @ Politico</title>
-		<link rel="stylesheet" type="text/css" href="public/css/style.css">
-	</head>
-	<body>
-		<div class="row">
-			<h1 id="h1-redirect-title" class="text-center">Go to authentication page in 3 seconds</h1>
-		</div>
-		<script type="text/javascript">
-			var time = setTimeout(function(){
-		        window.location='UI/signin.html';
-		    }, 3000);
-		</script>
-	</body>
-</html>
