+++ package.json
@@ -5,7 +5,7 @@
   "main": "app.js",
   "scripts": {
     "start": "node app.js",
-    "test": "nyc --reporter=html --reporter=text mocha --exit",
+    "test": "nyc --reporter=html --reporter=text mocha server/test/*.js --exit",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
@@ -34,16 +34,9 @@
   },
   "dependencies": {
     "body-parser": "^1.18.3",
-    "compression": "^1.7.3",
-    "consolidate": "^0.15.1",
-    "cookie-parser": "^1.4.3",
-    "ejs": "^2.6.1",
     "errorhandler": "^1.5.0",
     "express": "^4.16.4",
-    "express-session": "^1.15.6",
-    "express-validator": "^5.3.1",
-    "path": "^0.12.7",
-    "randomstring": "^1.1.5"
+    "express-validator": "^5.3.1"
   },
   "engines": {
     "node": "11.0.0",
