+++ server/helper/data/offices.json
@@ -1,0 +1,1 @@
+[{"id":7,"name":"Office Test","type":"state","createdAt":"Thu Feb 14 2019 00:26:25 GMT+0200 (Central Africa Time)","updatedAt":"Thu Feb 14 2019 00:26:25 GMT+0200 (Central Africa Time)"},{"id":8,"name":"Office4","type":"state","createdAt":"Thu Feb 14 2019 00:28:42 GMT+0200 (Central Africa Time)","updatedAt":"Thu Feb 14 2019 00:28:42 GMT+0200 (Central Africa Time)"},{"id":9,"name":"Office Test","type":"legistrative","createdAt":"Thu Feb 14 2019 00:29:57 GMT+0200 (Central Africa Time)","updatedAt":"Thu Feb 14 2019 00:29:57 GMT+0200 (Central Africa Time)"},{"id":10,"name":"Office Test","type":"federal","createdAt":"Thu Feb 14 2019 00:31:31 GMT+0200 (Central Africa Time)","updatedAt":"Thu Feb 14 2019 00:31:31 GMT+0200 (Central Africa Time)"}]
