+++ server/helper/data/parties.json
@@ -1,0 +1,1 @@
+[{"id":9,"name":"Part Test","hqAddress":"Kigali","logoUrl":"testUrl","createdAt":"Thu Feb 14 2019 00:31:31 GMT+0200 (Central Africa Time)","updatedAt":"Thu Feb 14 2019 00:31:31 GMT+0200 (Central Africa Time)"}]
